include icons in token theme
